refactor(arduino): move and organize Arduino code
