+++ .gitattributes
@@ -279,15 +279,17 @@ Procfile    text
 '*.apk' filter=lfs diff=lfs merge=lfs -text
 
 ## Linguist: custom DSL specification ##
-*.qa  linguist-language=Xtext
-*.qak linguist-language=Xtext
+*.qa  text linguist-language=Xtext linguist-detectable=true
+*.qak text linguist-language=Xtext linguist-detectable=true
 
 ## Linguist configurations ##
-*.fz  text linguist-language=xml
-src-gen/* linguist-generated=true linguist-detectable=true
-srcMore/* linguist-generated=true linguist-detectable=true
-sites/* linguist-generated=true linguist-documentation
-docker/* linguist-generated=true
-build_*.gradle linguist-generated=true linguist-detectable=true
+*.fz          text  linguist-language=xml
+*.czrc        text  linguist-language=json
+src-gen/*           linguist-generated=true linguist-detectable=true
+srcMore/*           linguist-generated=true linguist-detectable=true
+sites/*             linguist-generated=true linguist-documentation
+soffritti/*         linguist-vendored
+docker/*            linguist-generated=true
+build_*.gradle      linguist-generated=true linguist-detectable=true
 it.unibo.libs2018/* linguist-vendored
-report/* linguist-documentation
+report/*            linguist-documentation
