+++ .vscode/extensions.json
@@ -2,6 +2,7 @@
     "recommendations": [
         "EditorConfig.EditorConfig",
         "James-Yu.latex-workshop",
-        "KnisterPeter.vscode-commitizen"
+        "KnisterPeter.vscode-commitizen",
+        "platformio.platformio-ide"
     ]
 }
