--- NanoBotArduino/src/Ultrasonic.cpp
@@ -1,75 +1,0 @@
-/*
- * Ultrasonic.cpp
- *
- * Library for Ultrasonic Ranging Module in a minimalist way
- *
- * created 3 Apr 2014
- * by Erick Simões (github: @ErickSimoes | twitter: @AloErickSimoes)
- * modified 23 Jan 2017
- * by Erick Simões (github: @ErickSimoes | twitter: @AloErickSimoes)
- * modified 04 Mar 2017
- * by Erick Simões (github: @ErickSimoes | twitter: @AloErickSimoes)
- * modified 15 May 2017
- * by Eliot Lim    (github: @eliotlim)
- * modified 10 Jun 2018
- * by Erick Simões (github: @ErickSimoes | twitter: @AloErickSimoes)
- * modified 14 Jun 2018
- * by Otacilio Maia (github: @OtacilioN | linkedIn: in/otacilio)
- *
- * Released into the MIT License.
- */
-
-#if ARDUINO >= 100
-  #include <Arduino.h>
-#else
-  #include <WProgram.h>
-#endif
-
-#include "Ultrasonic.h"
-
-Ultrasonic::Ultrasonic(uint8_t trigPin, uint8_t echoPin, unsigned long timeOut) {
-  trig = trigPin;
-  echo = echoPin;
-  threePins = trig == echo ? true : false;
-  pinMode(trig, OUTPUT);
-  pinMode(echo, INPUT);
-  timeout = timeOut;
-}
-
-unsigned int Ultrasonic::timing() {
-  if (threePins)
-    pinMode(trig, OUTPUT);
-
-  digitalWrite(trig, LOW);
-  delayMicroseconds(2);
-  digitalWrite(trig, HIGH);
-  delayMicroseconds(10);
-  digitalWrite(trig, LOW);
-
-  if (threePins)
-    pinMode(trig, INPUT);
-  
-  previousMicros = micros();
-  while(!digitalRead(echo) && (micros() - previousMicros) <= timeout); // wait for the echo pin HIGH or timeout
-  previousMicros = micros();
-  while(digitalRead(echo)  && (micros() - previousMicros) <= timeout); // wait for the echo pin LOW or timeout
-
-  return micros() - previousMicros; // duration
-}
-
-/*
- * If the unit of measure is not passed as a parameter,
- * sby default, it will return the distance in centimeters.
- * To change the default, replace CM by INC.
- */
-unsigned int Ultrasonic::read(uint8_t und) {
-  return timing() / und / 2;  //distance by divisor
-}
-
-/*
- * This method is too verbal, so, it's deprecated.
- * Use read() instead.
- */
-unsigned int Ultrasonic::distanceRead(uint8_t und) {
-  return read(und);
-}
