+++ arduino/.vscode/extensions.json
