+++ arduino/platformio.ini
@@ -13,3 +13,7 @@ platform = atmelavr
 board = uno
 framework = arduino
 
+lib_deps =
+  1507@^3.0.0   ; Ultrasonic
+  6193@^1.0.0   ; Drive
+
