+++ arduino/src/actuators/L298N.h
@@ -1,8 +1,7 @@
 #ifndef L298N_h
 #define L298N_h
 
-#include "Arduino.h"
-
+#include <Arduino.h>
 
 typedef void (*CallBackFunction) ();
