+++ arduino/src/sensors/StabilizedProximitySensor.cpp
